requirements.txt
eyed3>=0.9.5
Pillow>=8.0.1
python-slugify>=0.1.3
